Cast recomputed probabilities back to the input dtype in backward
With L in fp32, S - L promoted P to fp32 while dO and V stayed bf16, so
the gradient matmuls saw mismatched operands. Exponentiate in fp32 for
accuracy, then cast down.

+++ cs336_systems/FlashAttention.py
@@ -13,7 +13,9 @@ def flash_backward(Q, K, V, O, dO, L, is_causal=False):
         k_idx = torch.arange(K.shape[1], device=Q.device)
         S = S + torch.where(q_idx[:, None] >= k_idx[None, :], 0.0, -1e6)
 
-    P  = torch.exp(S - L.unsqueeze(-1))
+    # Exponentiate in fp32 since L may be fp32, then match the input dtype so
+    # the matmuls below see consistent operands.
+    P  = torch.exp(S.float() - L.unsqueeze(-1)).to(Q.dtype)
     dV = torch.einsum("bqk,bqd->bkd", P, dO)
     dP = torch.einsum("bqd,bkd->bqk", dO, V)
     D  = (O * dO).sum(dim=-1)
